reorganize flow and multilines classes
