+++ SkeletonViewCore/Sources/Internal/SkeletonConfigs/SkeletonConfig.swift
