--- SkeletonViewCore/Sources/Multilines/UITextView+Multiline.swift
@@ -1,72 +1,0 @@
-// Copyright © 2018 SkeletonView. All rights reserved.
-
-import UIKit
-
-public extension UITextView {
-    @IBInspectable
-    var lastLineFillPercent: Int {
-        get { return lastLineFillingPercent }
-        set { lastLineFillingPercent = min(newValue, 100) }
-    }
-
-    @IBInspectable
-    var linesCornerRadius: Int {
-        get { return multilineCornerRadius }
-        set { multilineCornerRadius = newValue }
-    }
-
-    @IBInspectable
-    var skeletonLineSpacing: CGFloat {
-        get { return multilineSpacing }
-        set { multilineSpacing = newValue }
-    }
-
-    var skeletonPaddingInsets: UIEdgeInsets {
-        get { return paddingInsets }
-        set { paddingInsets = newValue }
-    }
-}
-
-extension UITextView: ContainsMultilineText {
-    var lineHeight: CGFloat {
-        if let fontLineHeight = font?.lineHeight {
-            if let heightConstraints = heightConstraints.first?.constant {
-                return (fontLineHeight > heightConstraints) ? heightConstraints : fontLineHeight
-            }
-            
-            return fontLineHeight
-        }
-        
-        return SkeletonAppearance.default.multilineHeight
-    }
-    
-    var numberOfLines: Int {
-        -1
-    }
-	
-    var lastLineFillingPercent: Int {
-        get {
-            let defaultValue = SkeletonAppearance.default.multilineLastLineFillPercent
-            return ao_get(pkey: &MultilineAssociatedKeys.lastLineFillingPercent) as? Int ?? defaultValue
-        }
-        set { ao_set(newValue, pkey: &MultilineAssociatedKeys.lastLineFillingPercent) }
-    }
-
-    var multilineCornerRadius: Int {
-        get {
-            let defaultValue = SkeletonAppearance.default.multilineCornerRadius
-            return ao_get(pkey: &MultilineAssociatedKeys.multilineCornerRadius) as? Int ?? defaultValue
-        }
-        set { ao_set(newValue, pkey: &MultilineAssociatedKeys.multilineCornerRadius) }
-    }
-
-    var multilineSpacing: CGFloat {
-        get { return ao_get(pkey: &MultilineAssociatedKeys.multilineSpacing) as? CGFloat ?? SkeletonAppearance.default.multilineSpacing }
-        set { ao_set(newValue, pkey: &MultilineAssociatedKeys.multilineSpacing) }
-    }
-
-    var paddingInsets: UIEdgeInsets {
-        get { return ao_get(pkey: &MultilineAssociatedKeys.paddingInsets) as? UIEdgeInsets ?? .zero }
-        set { ao_set(newValue, pkey: &MultilineAssociatedKeys.paddingInsets) }
-    }
-}
